pyproject.toml
[project]
name = "self-instruct"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "openpyxl>=3.1.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0",
    "rouge-score>=0.1.2",
    "tqdm>=4.66.0",
    "numpy>=1.24.0",
    "requests>=2.32.4",
]
